+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * `all_for_order_book/2` and `all_for_order_book/3` for listing trades for an order book
 - `Order Book`
   * `get/1` for getting order book details
+- `PaymentPaths`
+  * `all/0` and `all/1` for listing payment paths
 
 ## [0.1.1] - 2018-01-27
 ### Fixed